omop-cdm-v3/final_pyspark_omop/requirements.txt
# PySpark ETL for Synthea to OMOP CDM conversion
# Python requirements file

# Core PySpark
pyspark>=3.3.0,<4.0.0

# YAML configuration support
PyYAML>=6.0

# Optional: Database connectivity (uncomment as needed)
# For PostgreSQL
# psycopg2-binary>=2.9.0

# For SQL Server
# pyodbc>=4.0.0

# For MySQL
# PyMySQL>=1.0.0

# For Oracle
# cx_Oracle>=8.0.0

# Development and testing (optional)
# pytest>=7.0.0
# pytest-spark>=0.6.0

# Jupyter notebook support (optional)
# jupyter>=1.0.0
# notebook>=6.0.0

# Data quality and profiling (optional)
# pandas>=1.5.0
# great-expectations>=0.15.0
